third_party/blink/web_tests/svg/parser/whitespace-integer-expected.txt

CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47a".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "0a".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32a".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245a".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47e".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "0e".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32e".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245e".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47foo".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "0foo".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32foo".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245foo".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47)90".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "0)90".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32)90".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245)90".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47em".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "0em".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32em".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245em".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47ex".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "0ex".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32ex".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245ex".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47px".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "0px".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32px".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245px".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47in".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "0in".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32in".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245in".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47cm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "0cm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32cm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245cm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47mm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "0mm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32mm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245mm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47pt".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "0pt".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32pt".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245pt".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47pc".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "0pc".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32pc".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245pc".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47%".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "0%".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32%".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245%".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47a".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0a".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32a".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245a".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47e".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0e".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32e".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245e".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47foo".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0foo".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32foo".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245foo".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47)90".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0)90".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32)90".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245)90".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47em".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0em".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32em".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245em".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47ex".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0ex".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32ex".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245ex".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47px".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0px".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32px".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245px".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47in".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0in".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32in".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245in".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47cm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0cm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32cm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245cm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47mm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0mm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32mm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245mm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47pt".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0pt".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32pt".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245pt".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47pc".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0pc".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32pc".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245pc".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47%".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0%".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32%".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245%".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47a".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0a".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32a".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245a".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47e".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0e".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32e".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245e".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47foo".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0foo".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32foo".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245foo".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47)90".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0)90".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32)90".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245)90".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47em".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0em".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32em".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245em".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47ex".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0ex".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32ex".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245ex".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47px".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0px".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32px".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245px".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47in".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0in".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32in".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245in".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47cm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0cm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32cm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245cm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47mm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0mm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32mm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245mm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47pt".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0pt".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32pt".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245pt".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47pc".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0pc".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32pc".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245pc".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47%".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0%".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32%".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245%".
This is a testharness.js-based test.
Harness: the test ran to completion.
